requirements.txt
# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0
requests>=2.31.0
aiohttp>=3.9.0

# LLM clients
anthropic>=0.18.0
openai>=1.10.0

# Search & Content
arxiv>=2.0.0
beautifulsoup4>=4.12.0
trafilatura>=1.6.0

# Export
markdown>=3.5.0
python-docx>=1.1.0
weasyprint>=60.0

# Utilities
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.7.0
tiktoken>=0.5.0
